# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for ! boxfight shop Discord dashboard
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface CurrentUser {
  id: string;
  username: string;
  discriminator: string;
  /** @nullable */
  avatar?: string | null;
  isAdmin: boolean;
  isAuthorized: boolean;
  hasValidToken: boolean;
  /** @nullable */
  email?: string | null;
}

export interface AuthorizedUser {
  id: number;
  discordId: string;
  username: string;
  /** @nullable */
  avatar?: string | null;
  authorizedBy: string;
  authorizedAt: string;
  /** @nullable */
  notes?: string | null;
}

export interface AuthorizedUserInput {
  discordId: string;
  username?: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface StoredToken {
  id: number;
  discordUserId: string;
  discordUsername: string;
  /** @nullable */
  discordAvatar?: string | null;
  token: string;
  addedAt: string;
  addedByDiscordId: string;
  addedByUsername: string;
  /** @nullable */
  lastValidated?: string | null;
}

export interface TokenInput {
  token: string;
}

export interface TokenValidationResult {
  valid: boolean;
  /** @nullable */
  discordUserId?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  discriminator?: string | null;
  message: string;
}

export interface MessageInput {
  token: string;
  channelId: string;
  message: string;
  /** Initial delay in seconds before starting */
  delaySeconds: number;
  /** Interval between messages in seconds */
  intervalSeconds: number;
  /** How many times to send (0 = infinite) */
  repeatCount: number;
}

export type MessageJobStatus = typeof MessageJobStatus[keyof typeof MessageJobStatus];


export const MessageJobStatus = {
  pending: 'pending',
  running: 'running',
  stopped: 'stopped',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface MessageJob {
  id: string;
  channelId: string;
  message: string;
  delaySeconds: number;
  intervalSeconds: number;
  repeatCount: number;
  sentCount: number;
  status: MessageJobStatus;
  createdAt: string;
  /** @nullable */
  discordUserId?: string | null;
  /** @nullable */
  discordUsername?: string | null;
}

export interface DashboardStats {
  totalAuthorizedUsers: number;
  activeJobs: number;
  totalMessagesSent: number;
  totalTokens?: number;
  totalJobs?: number;
  pendingJobs?: number;
  stoppedJobs?: number;
  completedJobs?: number;
}

export type ListAuthorizedUsersParams = {
/**
 * Search query to filter users
 */
search?: string;
};