# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for ! boxfight shop Discord dashboard
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tokens/mine:
    get:
      operationId: listMyTokens
      tags: [tokens]
      summary: List tokens added by the current authenticated user
      responses:
        "200":
          description: List of stored tokens for current user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoredToken"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tokens:
    get:
      operationId: listTokens
      tags: [tokens]
      summary: List all stored user tokens with Discord user info (admin only)
      responses:
        "200":
          description: List of stored tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoredToken"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tokens/validate:
    post:
      operationId: validateToken
      tags: [tokens]
      summary: Validate a Discord user token and store it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenInput"
      responses:
        "200":
          description: Token validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenValidationResult"
        "400":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    StoredToken:
      type: object
      properties:
        id:
          type: integer
        discordUserId:
          type: string
        discordUsername:
          type: string
        discordAvatar:
          type: ["string", "null"]
        token:
          type: string
        addedAt:
          type: string
        addedByDiscordId:
          type: string
        addedByUsername:
          type: string
        lastValidated:
          type: ["string", "null"]
      required:
        - id
        - discordUserId
        - discordUsername
        - token
        - addedAt
        - addedByDiscordId
        - addedByUsername
    TokenInput:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    TokenValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        discordUserId:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        discriminator:
          type: ["string", "null"]
        message:
          type: string
      required:
        - valid
        - message
...